requirements.txt
Flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.5.4
Pillow==10.3.0
selenium==4.21.0
webdriver-manager==4.0.1
ImageHash==4.3
gunicorn==21.2.0
six

